+++ packages/react-dom/src/utils.ts
@@ -1,13 +1,10 @@
 import { MouseEvent, AnchorHTMLAttributes } from "react";
 
-export function canNavigate(
-  event: MouseEvent<HTMLElement>,
-  forward: AnchorHTMLAttributes<HTMLAnchorElement>
-) {
+export function canNavigate(event: MouseEvent<HTMLElement>, target?: string) {
   return (
     !event.defaultPrevented &&
     event.button === 0 &&
     !(event.metaKey || event.altKey || event.ctrlKey || event.shiftKey) &&
-    (!forward || !forward.target)
+    (!target || target !== "_self")
   );
 }