+++ packages/react-universal/src/hooks/useLocation.ts
@@ -1,5 +1,6 @@
 import React from "react";
 import useRouter from "./useRouter";
+import { stringifyFlatObject } from "../utils";
 
 import { RouteLocation } from "@curi/types";
 
@@ -7,10 +8,7 @@ export default function useLocation(props: RouteLocation) {
   const router = useRouter();
   const pathname = React.useMemo(() => {
     return props.name ? router.route.pathname(props.name, props.params) : "";
-  }, [
-    props.name,
-    ...Object.keys(props.params || {}).map(key => props.params[key])
-  ]);
+  }, [props.name, stringifyFlatObject(props.params)]);
 
   return {
     pathname,