+++ packages/react-universal/src/utils.ts
@@ -1,0 +1,5 @@
+export function stringifyFlatObject(obj: { [key: string]: any } = {}) {
+  return Object.keys(obj)
+    .map(key => `${key}=${obj[key]}`)
+    .join("&");
+}