+++ packages/react-universal/types/utils.d.ts
@@ -1,0 +1,3 @@
+export declare function stringifyFlatObject(obj?: {
+    [key: string]: any;
+}): string;