+++ packages/react-universal/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* `useNavigationHandler` & `useStatefulNavigationHandler` re-use event handler callback when possible. They also take a `target` string instead of the full `forward` props object.
+
 ## 2.0.0-beta.6
 
 * Revert snake case. `create_router_component` is now `createRouterComponent`.